requirements.txt
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.2.0
Automat==22.10.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.40
botocore==1.34.40
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.5
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.12.1
Deprecated==1.2.14
distro==1.9.0
Django==5.0.2
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-filter==23.5
django-phonenumber-field==8.0.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
entrypoints==0.4
et-xmlfile==1.1.0
Faker==25.0.1
favicon==0.7.0
filelock==3.13.4
fonttools==4.51.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
holidays==0.53
hpack==4.0.0
htbuilder==0.6.2
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.0
incremental==22.10.0
isodate==0.6.1
itemadapter==0.8.0
itemloaders==1.1.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
microsoft-kiota-abstractions==1.3.2
microsoft-kiota-authentication-azure==1.0.0
microsoft-kiota-http==1.3.1
microsoft-kiota-serialization-form==0.1.0
microsoft-kiota-serialization-json==1.2.0
microsoft-kiota-serialization-multipart==0.1.0
microsoft-kiota-serialization-text==1.0.0
more-itertools==10.2.0
msal==1.28.0
msal-extensions==1.1.0
msgraph-core==1.0.0
msrest==0.7.1
multidict==6.0.5
mysql-connector-python==8.4.0
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
openai==1.19.0
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.0
parsel==1.9.1
pdf2image==1.17.0
pendulum==3.0.0
phonenumbers==8.13.40
pillow==10.3.0
plotly==5.22.0
portalocker==2.8.2
prometheus_client==0.20.0
Protego==0.3.1
protobuf==4.25.3
pyarrow==15.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.9.0
PyDispatcher==2.0.7
pydyf==0.11.0
Pygments==2.18.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pymdown-extensions==10.8.1
pymongo==4.8.0
PyMySQL==1.1.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==4.3.1
pyphen==0.16.0
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.6.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
qrcode==7.4.2
queuelib==1.6.2
referencing==0.35.1
reportlab==4.2.2
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.0
scipy==1.12.0
Scrapy==2.11.1
selenium==4.19.0
service-identity==24.1.0
setuptools==69.2.0
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snscrape==0.7.0.20230622
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
st-annotated-text==4.0.1
std-uritemplate==0.0.57
streamlit==1.34.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
svglib==1.5.1
tenacity==8.3.0
time-machine==2.14.1
tinycss2==1.3.0
tldextract==5.1.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.0.7
validators==0.28.1
w3lib==2.1.2
watchdog==4.0.0
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.2
whitenoise==6.6.0
wrapt==1.16.0
wsproto==1.2.0
xhtml2pdf==0.2.16
yarl==1.9.4
zipp==3.18.1
zope.interface==6.2
zopfli==0.2.3